Refactor customFunctions and add afterFn (wip)
