+++ back/src/customFunctions/afterFn/index.ts
@@ -1,0 +1,13 @@
+import type { BasicPrivilegeProps, TransText } from '../../dbModels/_modelTypePartials';
+import type { UserData } from '../../utils/userAndPrivilegeChecks';
+
+type CustomAfterFn = {
+  [key: string]: {
+    name: TransText;
+    description: TransText;
+    privileges?: BasicPrivilegeProps;
+    afterFn: (data: unknown, userData: UserData) => boolean;
+  };
+};
+
+export const customValidators: CustomAfterFn = {};
